inode handling added.
On filesystems that limit inodes, this will attempt to ensure that enough are free to transfer files. Note that we still have to
handle duplicating all files in backup.0, which is going to be a big overhead.

Possible method - increase $reqinodes by `find backup.0 -printf "%i\n" | sort -u | wc -l`.